requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.2.26
langchain-openai==1.1.12
langchain-groq==1.1.2
langchain-tavily==0.2.17

langgraph==1.1.6
langgraph-checkpoint==4.0.1

tavily-python==0.7.23

pydantic==2.12.5
python-dotenv==1.2.2
requests==2.33.1

# image generation
pollinations.ai==4.5.1